Update smallPokedex at all
Updated and Added:
css/pageStyle.css and css/styles.css,
Added css classes and styles for the layout of the new added JS modal.
Updated and added pokemon-Button-Style classes, to add the specific color to the background of each pokemon modal, after checking the type of the pokemon. For 18 different pokemon types add 18 different color to the pokemon button.
Added media classes for display behaviour of the app, when screen size is getting smaller.

Updated and Added:
index.html,
updated and added div-tag and further buttons related to the creation of JS modal and some css classes to tags and elements.

Updated and Added:
js/scripts.js,
updated functions and changed function positions.
Added new function to show up details of the Pokemon from pokemon API and display them in a modal after clicking the mane button on the page .
Updated and added pokemon-Button-Style function, to add the specific color to the background of each pokemon modal, after checking the type of the pokemon. For 18 different pokemon types add 18 different color to the pokemon button.

Related: Intro to JavaScript; Task 1.8