Update smallPokedex at all
Added Bootstrap to the project (both the CSS and JS part).

Updated and Added:
index.html,
updated and added the proper uns of Bootstrap utility classes in the application,
use list-group for the ul element and group-list-item on li elements.
Replaced the modal in “index.html” with the Bootstrap modal.
Replace the top navigation bar with Bootstrap’s navbar components.
Made all components, such as the modal, navigation bar, and so on, are accessible to screen readers.

Removed:
css/pageStyle.css, at all.

Updated and Removed:
css/styles.css and js/scripts.js,
removed any old CSS and JavaScript code which is not need anymore, to follow and practice the good coding rules and keep the code clean and accessabel.

Related: Intro to JavaScript; Task 1.10